requirements.txt
# Excel handling
openpyxl>=3.1.5

# PDF processing
pdfplumber>=0.11.0

# LLM integration
openai>=1.50.0
python-dotenv>=1.0.0

# Optional: For better data handling
typing-extensions>=4.12.0